pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-doc-parser"
version = "0.1.0"
description = "使用大模型解析API文档的智能工具"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "API Doc Parser Team"}
]
keywords = ["api", "document", "parser", "llm", "openai", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web框架
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # CLI框架
    "typer>=0.9.0",
    "rich>=13.7.0",
    
    # 数据验证和配置
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # LLM SDK
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    
    # 文档处理
    "pymupdf>=1.23.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pandas>=2.1.0",
    
    # 任务队列
    "celery>=5.3.0",
    "redis>=5.0.0",
    
    # 工具和工具
    "tiktoken>=0.5.0",
    "structlog>=24.1.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
api-doc-parser = "src.cli:app"

[project.urls]
Homepage = "https://github.com/example/api-doc-parser"
Repository = "https://github.com/example/api-doc-parser"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
